setup-environment: make sure /bin/sh is actually /bin/bash
setup-environment cannot work, if /bin/sh is actually /bin/dash or an
actual bourne shell, since those won't pass the positional parameters
specified on the command line to the source'd script, but the
positional parameters that were passed to the original (login) shell
(which are usually empty).

Check for the BASH-specific environment variable BASH_VERSION, to make
sure, the underlying shell is actually 'bash'.