drm/panel: simple: make it possible to override LCD bus format
The baseboards for the Ka-Ro electronics series of i.MX modules have
their display ports wired to the topmost bits of the R,G,B bit
lanes. Thus no matter what LCD bus width (18 or 24bit) a module
provides, the LCD port can be used without shuffling bit lanes.

Add a function to handle certain quirks of the LCD interface to the
panel driver to be able to override the bus format specified in a
panel's display_mode.