linux-karo: Fix boot failure on QSMP-1351
When the stm32mp specific patches were applied to the Linux sourcecode
the patch stm32mp157-kconfig-dep-bugfix.patch, that fixed some Kconfig
dependencies for the stm32mp13 arch, was accidentally omitted.
This resulted in the QSMP-1351 not being able to boot Linux.

Switch to a kernel SRCREV that contains the above mentioned patch to
fiy this.